requirements.txt
Django==1.8.2
django-cors-headers==1.1.0
django-mptt==0.7.4
django-taggit==0.14.0
django-taggit-serializer==0.1.1
djangorestframework==3.1.2
djoser==0.3.0
